CouponCodeResponseDto 삭제 및 CouponResponseDto 변경
